Create skeleton screen and add shimmer
Create skeleton screen based on the original layout of the detail page, that is, create a new layout that mimics the original layout. This new layout will have blocks where normally the content would be shown, these blocks follow the specifications of the original elements such as margin, width and height.
The shimmer is an animation that serves to indicate the effect of loading.