Fix image scrollbars
With 'overflow: hidden' when passing the images through the buttons the images of the extremes presented a bug, in the lateral movement   some pieces of the images remained, sometimes outside the limits and sometimes within the limits of the grid - or scroll. With 'overflow-x: scroll' and hiding it but maintaining the functionality, the images pass normally.