Create authentication context
Create a login function that will do the authentication, add 'authorization' to the request header with the value "Bearer {token}" and set to a state that the user is logged in.
Provide the authentication function and the status that shows that the user is logged in for the entire application.