Create the pendency page to orphanages
Use the orphanage creation page as the basis for creating it, but blocking the data so that it is only for viewing.
Add a button to confirm the orphanage and remove it from the list of pending orphanages and a button to cancel the orphanage by deleting it from the database.