Create the forgot page
Search in the backend if you have a registered user with the provided email. If so, send an email to reset the password.
When sending the e-mail the send button will be changed to another indicating that the e-mail was sent and will be disabled, leaving only the option to return.