Update backend in relation to images
I had placed the API to save the images in the "public/uploads folder on the frontend for quick access to the images and for having a security problem when accessing local files. Later there was also another problem: when submitting the form with images, the API uploaded the image in the public folder of the frontend, but some folders within the react-app project including the 'public' were under hot reloading from the webpack DevServer, so each when saving a set of images, the browser reloaded the page.
Now I will use the express.static middleware function that is integrated with Express that allows to deliver static files such as images. So to access the API photos on the frontend, just use "http://localhost:3333/uploads/name-image" in the src property of the img tag.