Create the reset password page
In the email received, there will be a link to an application route using query params identifying a token. Decoding the token will have the email of the user who made the 'I forgot my password' request.
The user must enter the new password and confirmation, if it passes the validation criteria, the data will be sent to the API to update the password.
When sending data the 'Trocar senha' button will be changed to another one that contains a 'loading' indicating that the operation is in progress and when the operation is completed a 'check' will be shown indicating that the operation has completed.